+++ frontend/src/setupTests.js
@@ -7,13 +7,12 @@ vi.mock('@mui/material', async () => {
     ...actual,
     Menu: ({ open, children }) =>
       open ? React.createElement(actual.MenuList, { 'data-testid': 'menu' }, children) : null,
-    TextField: ({ inputRef, InputProps, inputProps, ...props }) =>
+    TextField: ({ inputRef, slotProps, ...props }) =>
       React.createElement(actual.TextField, {
         ...props,
         autoFocus: false,
         inputRef,
-        InputProps,
-        inputProps,
+        slotProps,
       }),
   };
 });